% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto.R
\name{get_carto}
\alias{get_carto}
\title{Get data from the City's Carto API}
\usage{
get_carto(query, format = "csv",
  base_url = "https://phl.carto.com/api/v2/sql", stringsAsFactors = F,
  ...)
}
\arguments{
\item{query}{character, SQL query}

\item{format}{character, desired response format, either "csv" (returned as
data.frame) or "json" (parsed into list by jsonlite::fromJSON)}

\item{base_url}{character, API endpoint, defaults to Philadelphia's Carto SQL
endpoint}

\item{stringsAsFactors}{logical, defaults to FALSE, applies to result if
format = "csv"}

\item{...}{other arguments passed either to read.csv or fromJSON, depending
on format}
}
\description{
Philadelphia's open data is published to carto, which allows you to query the
data using SQL (specifically the PostgreSQL flavor with the PostGIS
extension). Rather than executing the queries in a database program, they are
executed through HTTP (ie. the web browser) via Carto's SQL API.
}
\examples{

query <- paste("SELECT service_name, COUNT(*) AS n",
               "FROM public_cases_fc",
               "GROUP BY service_name",
               "ORDER BY n DESC")

get_carto(query)
get_carto(query, format = "json")

}
