% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_colors}
\alias{get_colors}
\title{Get names and hex codes from City of Philadelphia color palette}
\usage{
get_colors(..., palette = phl_colors)
}
\arguments{
\item{...}{List of colors to return. If \code{NULL}, returns the entire set of options.}

\item{palette}{Named character vector of hex codes; defaults to \code{phl_standards}.}
}
\description{
Get names and hex codes from City of Philadelphia color palette
}
\examples{
get_colors()
get_colors("ben_franklin_blue")
get_colors("bell_yellow", "ben_franklin_blue")

}
