% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_plot.R
\name{marg_plot}
\alias{marg_plot}
\title{Function to plot marginal levels / effects}
\usage{
marg_plot(res, dat, y_scale = function(x) scales::percent(x, accuracy =
  0.1))
}
\arguments{
\item{res}{Object output from \code{modmarg::marg}}

\item{dat}{Dataset used to estimate models (used to recover sample sizes)}

\item{y_scale}{Function that takes a numeric argument and returns a character for output labels}
}
\description{
Function to plot marginal levels / effects
}
\examples{
library(dplyr)
library(ggplot2)
library(extrafont)

data(mtcars)
dat <- mtcars \%>\%
  mutate_at(vars(cyl, am), as.factor)

mod <- glm(mpg ~ disp + cyl, data = dat)
res <- modmarg::marg(mod, var_interest = "cyl")
print(res)

marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

mod_sub <- glm(mpg ~ disp + cyl * am, data = dat)
res <- sub_marg(mod_sub, var_interest = "cyl", subgroup = "am")
print(res)

marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

res <- sub_marg(mod_sub, var_interest = "cyl", subgroup = "am",
                type = "effects")
print(res)

marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

}
