% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_tables.R
\name{marg_table}
\alias{marg_table}
\alias{marg_sub_table}
\title{Function to produce tables of marginal levels / effects}
\usage{
marg_table(lev, eff)

marg_sub_table(lev, eff)
}
\arguments{
\item{lev}{output of \code{modmarg::marg} with \code{type = "levels"}}

\item{eff}{output of \code{modmarg::marg} with \code{type = "effects"}}
}
\description{
Function to produce tables of marginal levels / effects

\code{marg_sub_table} handles output from \code{sub_marg}
}
\examples{
library(dplyr)
library(ggplot2)
library(extrafont)

data(mtcars)
dat <- mtcars \%>\%
  mutate(cyl = as.factor(cyl))
mod <- glm(mpg ~ disp + cyl, data = dat)
lev <- modmarg::marg(mod, var_interest = "cyl")
eff <- modmarg::marg(mod, var_interest = "cyl", type = "effects")

marg_table(lev, eff)


data(mtcars)
dat <- mtcars \%>\%
  mutate_at(vars(cyl, am), as.factor)

mod <- glm(mpg ~ disp + cyl * am, data = dat)
sub_marg(
  mod = mod,
  subgroup = "am",
  var_interest = "cyl",
  cofint = 0.9)
}
