% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{sub_marg}
\alias{sub_marg}
\title{Function to calculate margins for subgroups}
\usage{
sub_marg(mod, subgroup, var_interest = "treat", cofint = 0.9, ...)
}
\arguments{
\item{mod}{Any model object compatible with modmarg::marg (currently glm and
ivreg)}

\item{subgroup}{character, name of the variable to group by}

\item{var_interest}{character,  name of the treatment variable}

\item{cofint}{numeric, desired confidence level, must be in (0, 1)}

\item{...}{other arguments passed to modmarg::marg}
}
\description{
This function calculates marginal treatment effects for sub-populations, from
model objects compatible with \code{modmarg::marg}
}
