% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_plot.R
\name{marg_plot}
\alias{marg_plot}
\title{Function to plot marginal levels / effects}
\usage{
marg_plot(res, dat, y_scale = function(x) scales::percent(x, accuracy =
  0.1), color_fun = scale_color_phl, label_size = 3)
}
\arguments{
\item{res}{Object output from \code{modmarg::marg}}

\item{dat}{Dataset used to estimate models (used to recover sample sizes)}

\item{y_scale}{Function that takes a numeric argument and returns a character
for output labels}
}
\description{
In most situations, the plot will place treatments on the x axis, predicted
values (whether levels or effects) on the y axis, and will facet by subgroup,
if any. The only exception is when the input is marginal *effects* rather than
marginal *levels*, **and** there are only two treatment levels. In this case,
the x axis will display subgroup, since there would be only one treatment to
display in each facet.
}
\examples{
library(dplyr)
library(ggplot2)
library(extrafont)

data(mtcars)
dat <- mtcars \%>\%
  mutate_at(vars(cyl, am), as.factor)

# Overall
mod <- glm(mpg ~ disp + cyl, data = dat)

# Main levels
res <- modmarg::marg(mod, var_interest = "cyl")
marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

# Main effects
res <- modmarg::marg(mod, var_interest = "cyl", type = "effects")
marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

# Subgroups
mod_sub <- glm(mpg ~ disp + cyl * am, data = dat)

# Levels
res <- sub_marg(mod_sub, var_interest = "cyl", subgroup = "am")
marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

# Effects
res <- sub_marg(mod_sub, var_interest = "cyl", subgroup = "am",
                type = "effects")
marg_plot(
  res,
  dat = dat,
  y_scale = function(x) round(x, digits = 1)) +
  labs(x = "Cylinders", y = "Predicted miles per gallon")

# Subgroup effects with two treatment levels
res <- sub_marg(
  mod_sub,
  var_interest = "am",
  subgroup = "cyl",
  type = "effects")

marg_plot(
res, dat = dat, y_scale = function(x) round(x, digits = 1)) +
labs(x = "am", y = "Predicted miles per gallon")

}
